requirements.txt
pyserial
ble-serial
pygame
